Move RobotPoolHelper.h templates to unnamed namespace
